+++ sdk/src/org.graalvm.polyglot.tck/src/org/graalvm/polyglot/tck/LanguageProvider.java
@@ -96,7 +96,7 @@ public interface LanguageProvider {
      *
      * @param context the context for a guest language code literal evaluation
      * @return the {@link Snippet} representing the identity function
-     * @since 1.0
+     * @since 20.0.0 beta 1
      */
     default Snippet createIdentityFunctionSnippet(Context context) {
         Value value = createIdentityFunction(context);