+++ sdk/src/org.graalvm.polyglot/src/org/graalvm/polyglot/io/FileSystem.java
@@ -40,6 +40,7 @@
  */
 package org.graalvm.polyglot.io;
 
+import java.io.File;
 import java.io.IOException;
 import java.net.URI;
 import java.nio.channels.SeekableByteChannel;
@@ -347,6 +348,17 @@ public interface FileSystem {
      */
     default String getSeparator() {
         return parsePath("").getFileSystem().getSeparator();
+    }
+
+    /**
+     * Returns the path separator used to separate filenames in a path list. On UNIX the path
+     * separator is {@code ':'}. On Windows it's {@code ';'}.
+     *
+     * @return the path separator
+     * @since 20.0.0 beta 1
+     */
+    default String getPathSeparator() {
+        return File.pathSeparator;
     }
 
     /**